pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "feishu-agent"
version = "0.1.0"
description = "Local Feishu control agent powered by Ark Responses API and lark-cli."
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115,<1.0",
  "openai>=1.76.0,<2.0",
  "pydantic>=2.8,<3.0",
  "python-dotenv>=1.0,<2.0",
  "uvicorn>=0.30,<1.0",
  "httpx>=0.27,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
]

[project.scripts]
feishu-agent = "feishu_agent.app:run"
feishu-agent-shell = "feishu_agent.shell:run"
feishu-agent-reply-bot = "feishu_agent.auto_reply:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
